requirements.txt
aiohttp==3.6.2
appnope==0.1.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.3.0
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
Faker==4.0.2
filelock==3.0.12
funcsigs==1.0.2
google==2.0.3
grpcio==1.27.2
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
more-itertools==8.2.0
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.34
packaging==20.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.0
pdf2image==1.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
py==1.8.1
py-spy==0.3.3
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.4.1
requests==2.23.0
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
tabulate==0.8.6
tensorboardX==2.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
torch==1.4.0
torchvision==0.5.0
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
yarl==1.4.2
zipp==2.1.0
